rename githubUtils file
